added java script to provide form functionality
First I brought the focus to the "name" field.

I also added code and listener to hide the "other" job  and listen for when "other" is chosen, then show the appropriate field.

I next added code to listen for changes on the Tshirt "design" element and displays the appropriate theme for a given design.

last I added coded to listen for the activities checkboxes to be checked.  Once chosen, the total cost of
 the conference  is totaled up at the bottom.